requirements.txt
streamlit
langchain
langchain-openai
langchain-community
openai
faiss-cpu
python-dotenv
sentence-transformers
